Flatten Harness into .testing
