+++ tests/conftest.py
@@ -1,4 +1,4 @@
-from aspen.testing.harness import teardown
+from aspen.testing import teardown
 from aspen.testing.pytest_fixtures import client, harness, fs, website
 from aspen.testing.pytest_fixtures import sys_path, sys_path_scrubber
